Changed message inbox styling
Not sure which is the better option here - changes made to inbox messages but I've left the sent messages as default for now until team have a chance to review and decide what they think is best.

Note: Table striping is lost when we use a separate row to display the message header on a second row.